# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ECG POS API",
        "description": "Production REST API for JWT authentication, OIRS terminal validation, payment notification, and invoice generation.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/api/v1/validate-ipn": {
            "post": {
                "tags": [
                    "OIRS"
                ],
                "summary": "Validate IPN",
                "description": "Validate an OIRS IPN for a POS terminal. Authentication: Bearer token required. Validation rules: ipn is required string max 50; terminal_id is required string max 50.",
                "operationId": "validateIpn",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValidateIpnRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "IPN validated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentValidationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/payment-notification": {
            "post": {
                "tags": [
                    "OIRS"
                ],
                "summary": "Send payment notification",
                "description": "Notify OIRS about a successful terminal payment. Authentication: Bearer token required. Validation rules: ipn required string max 50; amount_paid required numeric greater than zero; terminal_id required string max 50; paid_at required date.",
                "operationId": "paymentNotification",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentNotificationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment notification sent successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentNotificationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/invoices": {
            "post": {
                "tags": [
                    "OIRS"
                ],
                "summary": "Generate invoice",
                "description": "Generate an OIRS invoice and authorization URL. Authentication: Bearer token required. Validation rules: revenue_code required string max 50; agency_code required string max 50; amount required numeric greater than zero; pid required string max 50; payment_type required and must be individual or corporate.",
                "operationId": "generateInvoice",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenerateInvoiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice generated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ValidateIpnRequest": {
                "required": [
                    "ipn",
                    "terminal_id"
                ],
                "properties": {
                    "ipn": {
                        "type": "string",
                        "example": "931713074597",
                        "maxLength": 50
                    },
                    "terminal_id": {
                        "type": "string",
                        "example": "1234567890",
                        "maxLength": 50
                    }
                },
                "type": "object"
            },
            "PaymentValidationResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "ipn": {
                                "type": "string",
                                "example": "931713074597"
                            },
                            "terminal_id": {
                                "type": "string",
                                "example": "1234567890"
                            },
                            "customer": {
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "79453121",
                                        "nullable": true
                                    },
                                    "ipn": {
                                        "type": "string",
                                        "example": "931713074597",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08030000000",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Osogbo, Osun State",
                                        "nullable": true
                                    }
                                },
                                "type": "object",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PaymentNotificationRequest": {
                "required": [
                    "ipn",
                    "amount_paid",
                    "terminal_id",
                    "paid_at"
                ],
                "properties": {
                    "ipn": {
                        "type": "string",
                        "example": "931713074597",
                        "maxLength": 50
                    },
                    "amount_paid": {
                        "type": "number",
                        "format": "float",
                        "example": 100,
                        "minimum": 0.01
                    },
                    "terminal_id": {
                        "type": "string",
                        "example": "1234567890",
                        "maxLength": 50
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-06 10:30:00"
                    }
                },
                "type": "object"
            },
            "PaymentNotificationResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "ipn": {
                                "type": "string",
                                "example": "931713074597"
                            },
                            "amount_paid": {
                                "type": "number",
                                "format": "float",
                                "example": 100
                            },
                            "terminal_id": {
                                "type": "string",
                                "example": "1234567890"
                            },
                            "paid_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-07-06T10:30:00+01:00"
                            },
                            "reference": {
                                "type": "string",
                                "example": "TRX-123456",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "GenerateInvoiceRequest": {
                "required": [
                    "revenue_code",
                    "agency_code",
                    "amount",
                    "pid",
                    "payment_type"
                ],
                "properties": {
                    "revenue_code": {
                        "type": "string",
                        "example": "4020154",
                        "maxLength": 50
                    },
                    "agency_code": {
                        "type": "string",
                        "example": "4100000",
                        "maxLength": 50
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 100,
                        "minimum": 0.01
                    },
                    "pid": {
                        "type": "string",
                        "example": "79453121",
                        "maxLength": 50
                    },
                    "payment_type": {
                        "type": "string",
                        "example": "individual",
                        "enum": [
                            "individual",
                            "corporate"
                        ]
                    }
                },
                "type": "object"
            },
            "InvoiceResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "ipn": {
                                "type": "string",
                                "example": "873648221564"
                            },
                            "authorization_url": {
                                "type": "string",
                                "format": "uri",
                                "example": "https://erms.cloudware.ng/gateway/checkout/873648221564/checkout"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token in the format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}